Created Interface for NodeIPPortDetector and a basic Unit test
