+++ src/freenet/node/ProtectedNodeIPPortDetector.java
@@ -1,0 +1,12 @@
+package freenet.node;
+
+import freenet.io.comm.Peer;
+
+interface ProtectedNodeIPPortDetector extends NodeIPPortDetector {
+
+    Peer[] detectPrimaryPeers();
+
+    void update();
+
+    void startARK();
+}
